% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_merip.R
\name{DESeq2_merip}
\alias{DESeq2_merip}
\title{Analysis MeRIP datasets with DESeq2.}
\usage{
DESeq2_merip(SE_M, MODE = "Meth", PCA = FALSE, HDER = "Unknown")
}
\arguments{
\item{SE_M}{A \code{SummarizedExperiment} object with 1 necessary collumn in \code{colData}: c("IP_input").}

\item{MODE}{Could be either "Meth" or "DM", the later will conduct differential methylation analysis with the design:}

\item{PCA}{Wheather to plot a PCA after an rlog transformation, default is FALSE.}

\item{HDER}{What should be the header of the PCA plot, applied when \code{PCA} = TRUE.

log2(Q) = intercept + I(Treated) + I(IP) + I(IP):I(Treated).

The result is just the differences in conditioning effect, or the statistics for estimate before the term I(IP):I(Treated).

We don't get this by contrast, because the information is already contained in coefficient estimate under the design above, and we don't need to linear combine (or a linear combination by t(c(0,0,0,1))) the estimates to get it.

An additional collumn c("Perturbation") is necessary for this option, the Perturbation collumn has to include character "C" for control condition.}
}
\value{
A DESeq2 result object.
}
\description{
\code{DESeq2_merip} is an internal function used to infer methylation and differential methylation given merip datasets.
}
