% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference_merip.R
\name{Inference_merip}
\alias{Inference_merip}
\title{Analysis MeRIP datasets with DESeq2.}
\usage{
Inference_merip(SE_M, MODE = "Meth", DM_METHOD = "DESeq2", PCA = FALSE,
  HDER = "Unknown", ROW_FILTER = 0, CQN = FALSE, GC_INDX = NULL)
}
\arguments{
\item{SE_M}{A \code{SummarizedExperiment} object with 1 necessary column in \code{colData} named "IP_input", its content should be a character vector consists of 2 values: c("IP", "input").}

\item{MODE}{Could be either "Meth" or "DM", the later will conduct differential methylation analysis with the design:}

\item{DM_METHOD}{A character string indicating the statistical method used in differential methylation analysis, could be one in c("DESeq2","QNB").}

\item{PCA}{Wheather to save the PCA plot after rlog transformation, default is FALSE, the plot will not be generated when \code{DM_METHOD} = "QNB" while the \code{MODE} = "DM".}

\item{HDER}{What should be the header of the PCA plot, applied when \code{PCA} = TRUE.}

\item{CQN}{Wheather to normalize GC content dependency of methylation / differential methylation log2FC, default is FALSE.}

\item{GC_INDX}{The GC content values for each features (rows) of the count matrix, it is required when CQN = TRUE.

log2(Q) = intercept + I(Treated) + I(IP) + I(IP):I(Treated).

The result is just the differences in conditioning effect, or the statistics for estimate before the term I(IP):I(Treated).

We don't get this by contrast, because the information is already contained in coefficient estimate under the design above, and we don't need to linear combine (or a linear combination by t(c(0,0,0,1))) the estimates to get it.

An additional column c("Perturbation") is necessary for this option, the Perturbation column has to include character "C" for control condition.}
}
\value{
A DESeq2 result object for DESeq2 analysis; for other analysis, it will generate a \code{data.frame} object.
}
\description{
\code{Inference_merip} is a function to infer methylation and differential methylation given merip datasets.
}
