% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_collumn_dist.R
\name{Plot_collumn_dist}
\alias{Plot_collumn_dist}
\title{Plot the collumn wise data distribution.}
\usage{
Plot_collumn_dist(M, TYPE = "box", HDER)
}
\arguments{
\item{M}{A \code{matrix}.}

\item{TYPE}{Can be either "density" or "box".}

\item{HDER}{The subtitle and the file name of the plot.}
}
\value{
A plot of the collumn wised density/boxplot.
}
\description{
\code{DESeq2_merip} is a function used to plot the collumned wise data distribution.
}
