% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_column_joint.R
\name{Plot_column_joint}
\alias{Plot_column_joint}
\title{Plot the joint distribution between matrix columns.}
\usage{
Plot_column_joint(M, METRIC = "euclidean", VISUAL = "MDS", HDER = "",
  GROUP_LABEL = NULL)
}
\arguments{
\item{M}{A \code{matrix}.}

\item{METRIC}{The metric used in clustering, can be one in "euclidean", "pearson", and "binary". The pearson metric is calculated as the euclidean metric after rescaling the columns.}

\item{HDER}{The subtitle and the file name of the plot.}

\item{GROUP_LABEL}{Optional, a vector used for colour labeling or faceting the dendrogram or MDS plot.}

\item{CLUSTER}{The clustering method, can be one in "MDS" and  "dendrogram".}
}
\value{
A plot for column wised clustering analysis.
}
\description{
\code{Plot_column_joint} is a function used to non-supervised learn / cluster the relationship between columns. (Insight into the joint distribution between samples)
}
\details{
By default, the column names of the matrix M will be used as the sample labels, other wise, it will use V_{1:ncol(M)}.
}
\examples{
Matrix_ex <- matrix(rnorm(9000),300,30)
Group_lab <- paste0( "V_",rep(1:10,each = 3) )
Plot_column_joint( Matrix_ex, "euclidean", "dendrogram", "Test1", Group_lab )
Plot_column_joint( Matrix_ex, "pearson", "MDS", "Test2", Group_lab )

}
\seealso{
\code{\link{Plot_column_marginal}}
}
