% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_ex_lengths.R
\name{Plot_ex_lengths}
\alias{Plot_ex_lengths}
\title{Plot exon length distribution plot based on the DESeq2 result.}
\usage{
Plot_ex_lengths(LIST_GR, TXDB, HDER)
}
\arguments{
\item{TXDB}{A \code{txdb} object.}

\item{HDER}{Determine the content of the title and the file name of the plot saved.}

\item{LIDT_GR}{A \code{list} of \code{GRanges}.}
}
\value{
A plot of the exon length distribution saved under the current working directory.
}
\description{
\code{DESeq2_merip} is an internal function used to infer methylation and differential methylation given merip datasets.
}
