% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GC_results.R
\name{Plot_GC_results}
\alias{Plot_GC_results}
\title{Plot GC content bias plot based on DESeq2 result.}
\usage{
Plot_GC_results(DS_RES, GC_IDX, HDER)
}
\arguments{
\item{DS_RES}{A \code{DESeqResults} object with additional factor column "Decision".}

\item{GC_IDX}{A numeric vector indicate the GC content of each individual feature.}

\item{HDER}{Determine the content of the title and the file name of the plot saved.}
}
\value{
A plot of the GC content saved under the current working directory.
}
\description{
\code{DESeq2_merip} is an internal function used to infer methylation and differential methylation given merip datasets.
}
