% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Seq_depth.R
\name{Plot_Seq_depth}
\alias{Plot_Seq_depth}
\title{Bar Plot of total sequencing depth.}
\usage{
Plot_Seq_depth(SE_M, HDER)
}
\arguments{
\item{SE_M}{A \code{SummarizedExperiment} object with 2 necessary collumns in \code{colData}: c( "IP_input", "SRR_RUN" ).}

\item{HDER}{Determine the content of the title and the file name of the plot saved.}
}
\value{
A pdf diagram saved under the current working directory.
}
\description{
\code{Plot_Seq_depth} is an internal function used to generate a sequencing depth plot given an input SummarizedExperiment object.
}
